requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.200       # YOLOv8

# Data
waymo-open-dataset-tf-2-12-0  # Waymo SDK
nuscenes-devkit>=1.1.11
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
pyarrow>=13.0.0

# MLOps
mlflow>=2.8.0
dvc>=3.30.0
dvc[gdrive]                # Google Drive remote

# Experiment Config
PyYAML>=6.0
omegaconf>=2.3.0

# Visualization & EDA
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
tqdm>=4.66.0

# Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
